requirements.txt
streamlit==1.44.1
pillow==11.1.0
tensorflow==2.16.1
numpy >= 1.26.0, <= 1.26.4
